synchronized() functions for Swift along with tests
This framework has three functions. One is supposed to be internal to the
framework but Xcode wants it to be public. The functions are:

 - `public func synchronized<T>(object: AnyObject, closure: () -> T) -> T`
 - `public func synchronized(object: AnyObject, closure: () -> Void)`
 - `void objc_synchronized(id object, void (^closure)());`

That last function is the one that should be internal so don't use it. The two
Swift functions are the same except one returns a value and the other doesn't.
Returning a value is useful if you want to write:

```
let value = synchronized(notThreadSafeDictionary) {
  notThreadSafeDictionary[key]
}
```