[breaking] Upgrade project for Xcode 12, change target deployment target to iOS 12
The project still runs on iOS 8 but Xcode 12 does not support it anymore and recommends setting the target iOS version to 12.0 and newer. This is a breaking change in the interface (tests will not ensure this library works on old versions of iOS) but not the implementation.

This also sets the Swift version to 5.4 and deletes the `.swift-version` file that CocoaPods has deprecated.